backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy>=2.0.38
psycopg[binary]>=3.2.0
groq==0.9.0
python-whois==0.9.4
dnspython==2.6.1
requests==2.32.3
httpx==0.27.0
certifi==2024.6.2
reportlab==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
slowapi==0.1.9
cryptography==42.0.8
pydantic>=2.10.0
pydantic-settings>=2.7.0
